+++ rotkehlchen/tests/api/test_settings.py
@@ -8,6 +8,7 @@ import requests
 
 from rotkehlchen.constants.assets import A_JPY
 from rotkehlchen.db.settings import (
+    DEFAULT_CURRENT_PRICE_ORACLES,
     DEFAULT_HISTORICAL_PRICE_ORACLES,
     DEFAULT_KRAKEN_ACCOUNT_TYPE,
     ROTKEHLCHEN_DB_VERSION,
@@ -98,6 +99,8 @@ def test_set_settings(rotkehlchen_api_server):
             value = ''
         elif setting == 'ksm_rpc_endpoint':
             value = 'http://kusama.node.com:9933'
+        elif setting == 'current_price_oracles':
+            value = ['coingecko', 'cryptocompare']
         elif setting == 'historical_price_oracles':
             value = ['coingecko', 'cryptocompare']
         else:
@@ -592,11 +595,17 @@ def test_queried_addresses_per_protocol(rotkehlchen_api_server):
     })
 
 
-def test_set_historical_price_oracles(rotkehlchen_api_server):
-    historical_price_oracles = list(DEFAULT_HISTORICAL_PRICE_ORACLES)
-    random.shuffle(historical_price_oracles)
+@pytest.mark.parametrize('setting, default_price_oracles', [
+    ('current_price_oracles', DEFAULT_CURRENT_PRICE_ORACLES),
+    ('historical_price_oracles', DEFAULT_HISTORICAL_PRICE_ORACLES),
+])
+def test_set_price_oracles(rotkehlchen_api_server, setting, default_price_oracles):
+    """Test the price oracles settings.
+    """
+    price_oracles = default_price_oracles.copy()
+    random.shuffle(price_oracles)
 
-    data = {'settings': {'historical_price_oracles': historical_price_oracles}}
+    data = {'settings': {setting: price_oracles}}
     response = requests.put(api_url_for(rotkehlchen_api_server, "settingsresource"), json=data)
     assert_proper_response(response)
 
@@ -604,9 +613,9 @@ def test_set_historical_price_oracles(rotkehlchen_api_server):
     json_data = response.json()
     result = json_data['result']
     assert json_data['message'] == ''
-    assert result['historical_price_oracles'] == historical_price_oracles
+    assert result[setting] == price_oracles
 
     # Check db settings
     rotki = rotkehlchen_api_server.rest_api.rotkehlchen
     settings = rotki.data.db.get_settings()
-    assert settings.historical_price_oracles == historical_price_oracles
+    assert getattr(settings, setting) == price_oracles